task 1

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -1,4 +1,3 @@
-#include "function_pointers.h"
 #include <stdio.h>
 #include <stdlib.h>
 
@@ -11,7 +10,7 @@
 
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-	unsigned int x;
+	size_t x;
 
 	if (array == NULL || action == NULL)
 		return;
